+++ src/mbgl/util/vec4.cpp
@@ -25,7 +25,7 @@
 using namespace mbgl;
 
 void matrix::transformMat4(vec4& out, vec4& a, mat4& m) {
-    float x = a[0], y = a[1], z = a[2], w = a[3];
+    double x = a[0], y = a[1], z = a[2], w = a[3];
     out[0] = m[0] * x + m[4] * y + m[8] * z + m[12] * w;
     out[1] = m[1] * x + m[5] * y + m[9] * z + m[13] * w;
     out[2] = m[2] * x + m[6] * y + m[10] * z + m[14] * w;
